Add annotations

+++ static/app.js
@@ -49,24 +49,77 @@ var App = React.createClass({
 });
 
 var ListItem = React.createClass({
+  getInitialState: function() {
+    return {
+      addingLinkAnnotation: false,
+      pendingLinkAnnotation: '',
+      linkAnnotations: this.props.data.link_annotations,
+    };
+  },
   delete: function() {
     this.props.deleteFn(this.props.data.id);
   },
+  toggleLinkAnnotationAdder: function() {
+    this.setState({addingLinkAnnotation: !this.state.addingLinkAnnotation});
+  },
+  pendingLinkAnnotationChanged: function(e) {
+    this.setState({pendingLinkAnnotation: e.target.value});
+  },
+  addLinkAnnotation: function() {
+    // TODO: The "link" string should live on the server
+    var annotation = {kind: "LINK", body: this.state.pendingLinkAnnotation};
+    var url = `/lists/${this.props.userId}/list/${this.props.listId}/items/${this.props.data.id}/annotations`;
+
+    console.log("POST(" + url + "): " + JSON.stringify(annotation));
+      
+    $.ajax({
+      url: url,
+      dataType: 'json',
+      type: 'POST',
+      data: JSON.stringify(annotation),
+      success: function(data) {
+        console.log("posted new annotation. got respnse: " + JSON.stringify(data));
+        // TODO: this assumes too much about the server objects and should live there
+        this.setState({linkAnnotations: this.state.linkAnnotations.concat(
+          [{url: data.body}]
+        )});
+      }.bind(this),
+      error: function(xhr, status, err) {
+        console.error(url, status, err.toString());
+      }.bind(this),
+    });
+    this.setState({addingLinkAnnotation: false, pendingLinkAnnotation: ''});
+  },
   render: function() {
-    var linkNodes = this.props.data.link_annotations.map(function(link) {
+    var linkNodes = this.state.linkAnnotations.map(function(link) {
       return (
-        <a href={link.url} key={link.url}>{link.url}</a>
+        <div key={link.url}>
+          <a href={link.url}>{link.url}</a>
+        </div>
       )
     });
+
+    var editNodes;
+    if (this.state.addingLinkAnnotation) {
+      editNodes =
+        <div>
+          <input type="text" placeholder="Url..." value={this.state.pendingLinkAnnotation} onChange={this.pendingLinkAnnotationChanged}/>
+          <button onClick={this.addLinkAnnotation}>+</button>
+        </div>;
+    }
 
     return (
       <li className="listItem">
         <div>
           <span className="name">{this.props.data.name}</span>
           <button onClick={this.delete}>X</button>
+          <button onClick={this.toggleLinkAnnotationAdder}>
+            {this.state.addingLinkAnnotation ? "-URL" : "+URL"}
+          </button>
         </div>
         <div className="description">{this.props.data.description}</div>
         {linkNodes}
+        {editNodes}
       </li>
     );
   }
@@ -172,7 +225,7 @@ var List = React.createClass({
   render: function() {
     var itemNodes = this.state.items.map(function(item) {
       return (
-        <ListItem data={item} key={item.id} deleteFn={this.deleteItem} />
+        <ListItem data={item} key={item.id} deleteFn={this.deleteItem} userId={this.props.params.userId} listId={this.props.params.listId}/>
       );
     }.bind(this));
     return (
